optimize build
- switch to native http
- fetch bombardier executable on demand
