+++ .gitignore
@@ -1,4 +1,4 @@
 .DS_Store
-/out
 
+bin
 node_modules
