+++ README.md
@@ -4,12 +4,13 @@ DDOS координатор, використовуючий bombardier на од
 
 ## Як використати
 
-Запуск через докер: `docker run --env SCHEDULED_TARGETS_URL=https://gloryforukraine.pages.dev/scheduled_targets.json PARALLEL_FACTOR=1 --rm -it archdev/gloryforukraine`
+Запуск через докер: `docker run --env SCHEDULED_TARGETS_URL=https://gloryforukraine.pages.dev/scheduled_targets.json PARALLEL_FACTOR=5 --rm -it archdev/gloryforukraine`
 
 Запуск через білд: 
 - install NodeJS 16+
 - git clone
 - cd gloryforukraine
+- npm i
 - `./scripts/launch.sh` or `.\scripts\launch.bat`
 
 # Як засетапи Gloryforukraine на Digigal Ocean App Platform
